Made the projects api more robust
This mostly revolved around adding better error handling. As well as checking if it belongs to the user which is a possible tempory solution. Assuming I get RLS to work between prisma and supabase. 

Additionally I chose to handle the errors in the api route over somewhere in the client files with the principle of declaring errors as close to their occurrence as possible